+++ siteweb/sae23/PHP/delcapteur.php
@@ -29,7 +29,7 @@
 						<?php } ?>
 						<?php
 							if($_SESSION['username'] != ""){ ?>
-								<li><a class="link" href="disconnect.php"><button>D&eacute;connexion</button></a></li>
+								<li><a class="link" href="disconnect.php">D&eacute;connexion</a></li>
 								<?php } else { ?>
 								<li><button id="logBtn" onclick="openPopup('logBtn','logPopup')">Connexion</button></li>
 						<?php } ?>
@@ -45,21 +45,37 @@
 				include ("mysql.php");
 				$nom_capteur = $_POST['nom_capteur'];
 				$type = $_POST['type'];
-				$nom_batiment = $_POST['nom_batiment']; 	
-				$requete = "DELETE FROM `capteur` WHERE `nom_capteur`='$nom_capteur' AND `type_capteur`='$type' AND `nom_batiment`='$nom_batiment'";
-				$resultat = mysqli_query($id_bd, $requete)
-					or die("Execution de la requete impossible : $requete");
-				mysqli_close($id_bd);
+				$nom_batiment = $_POST['nom_batiment']; 
+				$results = mysql_query("SELECT `nom_capteur` FROM `capteur` WHERE `nom_capteur`='$nom_capteur'; ");
+				$row = mysql_fetch_array($results);
+				if(!mysql_num_rows($results)){		
+					echo "<br /><strong>ce capteur existe pas, vérifier votre saisie SVP </strong><br />";
+						
+				} else {
+				
+					$requete = "DELETE FROM `capteur` WHERE `nom_capteur`='$nom_capteur' AND `type_capteur`='$type' AND `nom_batiment`='$nom_batiment'";
+					$resultat = mysqli_query($id_bd, $requete)
+						or die("Execution de la requete impossible : $requete");
+					$results = mysql_query("SELECT `nom_capteur` FROM `capteur` WHERE `nom_capteur`='$nom_capteur'; ");
+					$row = mysql_fetch_array($results);
+					if(!mysql_num_rows($results)){	
+						
+					
 
-				/* display the removed information */
-				echo '<div class="ajout">';
-				echo "<br /><strong>La donn&egravee suivante a &eacute;t&eacute; supprim&eacute;e au catalogue : </strong><br />";
-				echo "<ul>
-						<li> Nom du capteur : $nom_capteur</li>
-						<li> type du capteur : $type</li>
-						<li> Nom du batiment : $nom_batiment</li>
-						</ul>
-					</div>";	
+					/* display the removed information */
+					echo '<div class="ajout">';
+					echo "<br /><strong>La donn&egravee suivante a &eacute;t&eacute; supprim&eacute;e au catalogue : </strong><br />";
+					echo "<ul>
+							<li> Nom du capteur : $nom_capteur</li>
+							<li> type du capteur : $type</li>
+							<li> Nom du batiment : $nom_batiment</li>
+							</ul>
+						</div>";	
+					} else {	
+						echo "Vérifier votre saisie";
+						}
+					mysqli_close($id_bd);
+				}
 			?>
 			<hr />
 			<a href="../administration.php"><button>OK</button></a>
